+++ src/app/pages/ecosystem-admin/credits/send-credits/send-credits.component.ts
@@ -1,0 +1,20 @@
+import { Component, EventEmitter, Output } from '@angular/core';
+import { XcIconShoppingCartComponent, XcIconTimesComponent } from '@indziaki/ngx-xchange-icons';
+
+@Component({
+  selector: 'app-send-credits',
+  standalone: true,
+  imports: [
+    XcIconTimesComponent,
+    XcIconShoppingCartComponent
+  ],
+  templateUrl: './send-credits.component.html',
+  styleUrl: './send-credits.component.scss'
+})
+export class SendCreditsComponent {
+  @Output() close = new EventEmitter<void>();
+
+  onClose() {
+    this.close.emit();
+  }
+}